requirements.txt
asn1crypto==0.24.0
atomicwrites==1.2.1
attrs==18.2.0
azure-common==1.1.16
azure-nspkg==2.0.0
azure-storage==0.36.0
backports.functools-lru-cache==1.5
boto3==1.8.9
botocore==1.11.9
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
cryptography==2.3.1
docutils==0.14
enum34==1.1.6
funcsigs==1.0.2
future==0.16.0
futures==3.2.0
idna==2.7
ijson==2.3
ipaddress==1.0.22
jmespath==0.9.3
more-itertools==4.3.0
pathlib2==2.3.2
pluggy==0.8.0
py==1.7.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.19
pycryptodomex==3.6.6
PyJWT==1.6.4
pyOpenSSL==18.0.0
pytest==3.9.1
python-dateutil==2.7.3
pytz==2018.5
requests==2.20.0
s3transfer==0.1.13
scandir==1.9.0
six==1.11.0
snowflake-connector-python==1.6.10
urllib3==1.23
